[FIX] l10n_nl_reports: Closing moves state check
Before this commit:
As multiple closing moves can exist for a return, the check for the
state of the closing move was incorrect. It was assuming only one
closing move exists, which could lead to issues if multiple closing
moves were present.

After this commit:
The code now checks if any of the closing moves are in 'draft' state
before raising a warning.

opw-4976827

closes odoo/enterprise#91296

X-original-commit: 61265ba2489fff62b547c41f6fccf5475f983514
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>