[FIX] sale_planning: avoid generating shifts when creating SOL in shift
Before this commit, when the user clicks on add to last order in form
view of planning slot to add the slot into the last rental order. If no
SOL has been found inside that existing rental order, a new SOL will be
generated and linked to that slot. The problem is another shift will be
generated for the SOL once the SOL has been created.

This commit avoids generating the SOL when it is not needed.

task-4916650

Part-of: odoo/enterprise#93670
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>