[FIX] sign: redirect to correct website for company
According to a ticket, when you have 2 companies and send a sign
request from the second one, the button in the email received from the
signer redirects to the website of the first company. This may cause
confusion and may have legal implications. This PR aims to fix the
problem by correctly targeting the right website.

closes odoo/enterprise#91238

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>