[FIX] l10n_it_pos: skip receipt screen prints receipt
before this commit, if the option to skip the receipt screen
was ticked, then the receipt would never print.
This was caused by the printing logic being implemented on
the receipt screen instead of on the pos itself.
Furthermore, before this commit, the printReceipt logic in the module would not
pass the order to be printed.
This can become a problem upon context changes, where pos.get_order
does not return the completed order, but a newly created one.
This can for example happen when skipping the receipt screen with
the option to "print automatically" (iface_print_auto).

steps to reproduce:
1. install l10n_it_pos
2. configure one pos
3. configure the IT printer
4. select to skip the receipt screen (print automatically)
5. open the pos
6. make a sale
=> no ticket printed and the chrome console shows a printer error

With this new verison the printing logic was moved to the pos
so that printing of fiscal receipts with the italian fiscal printer
works, even when receipt screen is skipped.
This commit also adds the order as an argument, so we always print
the last completed order and not a newly created one.

opw-4882480

closes odoo/enterprise#96798

X-original-commit: 47413e4805068782cd924c0bc190bb300eeac2d6
Signed-off-by: Stéphane Vanmeerhaeghe (stva) <stva@odoo.com>
Signed-off-by: Victorien Sonneville (viso) <viso@odoo.com>