[FIX] hr_payroll: fix schedule pay not set
When installing the payroll app with a new db initialized in Australia, we would
 get a traceback because the schedule pay is False at some point.

This was due to the compute of schedule pay which did not handle the case where
the structure type is not set.

This fixes the issue by checking if the structure type is set to set the
schedule pay and adding a default value to monthly.

closes odoo/enterprise#91465

Ticket: 4974520
X-original-commit: 3404bdcd1d82648eae71f2fe3b163bc642b2b2d0
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>