[FIX] : wait before payment on discount
Before this commit, if a discount was applied with the blackbox, it
was applied after communication with blackbox which could be slow. If
the user was clicking payment before this disound was applied and had
only one payment method, a payment line with the old amount was added
which could lead to confusion and errors when this payment line was
sent to a terminal.
This is fixed by waiting for the discount to be applied before being
able to click on payment.

closes odoo/enterprise#94402

X-original-commit: 947bc27acdeb453444f5595683250dc7cb1f37b0
Related: odoo/odoo#226495
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>