[FIX] pos_iot: kitchen printers not working via websocket
Due to a missing `iot_id` field when the `DeviceController` for kitchen
printers is instantiated, the websocket fallback does not work. This
results in kitchen tickets failing to print even when regular receipts
are working correctly.

The fix is to provide the full IoT device model to the `DeviceController`
constructor, as is done elsewhere. To do so we now pass the `device_id`
field for the preparation printer model from the backend. We also now
send the relevant IoT devices if any IoT kitchen printers are
configured.

We also ensure IoT devices/boxes are loaded even with a 'limited' load.
This is because of an issue where an IoT device is added to the PoS
config, but because the device itself has not be written to recently,
the PoS doesn't load it and you get a traceback in the frontend.

closes odoo/enterprise#91464

X-original-commit: 4506682a3eb3a39f5f9f52590e6967029f78612a
Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>