[FIX] l10n_dk_{reports, instrastat}: rounding reports
This commit will change the options of the reports so that we have a rounding in
units since the intrastat and ec sales list report must be rounded.
This solution still allows people to have the report with decimals if needed in
the interface.

However, export file to be sent to the government like csv for ec sale list and
xlsx(idep) are now always rounded

task-4948271

closes odoo/enterprise#93868

X-original-commit: 28e78b6ba7080da3379ef7fb7ab5ebdac1db490f
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>