[IMP] appointment: support confirmation percentage without managing capacity
Before, we did not handle the manual_confirmation_percentage when
manage_capacity was False. Instead, if auto_confirm was True, meetings
were always 'booked' and 'request' otherwise.

However, we also have a quantity on the amount of meetings when
manage_capacity is False, e.g. the number of simultaneous meetings
per user / per resource for each slot.

Therefore, we now always display the percentage, even when manage_capacity
is False. The percentage is applied on the TOTAL CAPACITY, meaning the product
of the number of users / resources and of the max_bookings value.
When manage_capacity is True, it does not change, it is still the total of
resource capacity or the product of number of users and of user_capacity for
user-based appointments.

A test is modified to include this new behavior.

Task-4898893

Part-of: odoo/enterprise#88637
Related: odoo/upgrade#8014
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>