[FIX] sale_planning: hide to plan stat button when nothing to plan
Before this commit, when rental order contains 2 shifts for a whole day
and only one day is ordered, `To plan` stat button displayed in the form
view of a Sales/Rental order will show negative hours instead of being
hidden.

This commit makes sure that stat button is hidden when To plan is not
higher than 0 hour.

task-4916650

Part-of: odoo/enterprise#93670
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>