[IMP] sign,sign_itsme: Added QuickSign flow
Introduced a new functionality which allows users signing documents that
only contain autofillable information to skip the document filling step.
Currrently, even if fields are autofillable, users still need to click
on each field for the value to appear, then click on Validate and later
on Download.

With this PR, under the conditions that all fields are autofillable,
that there is only 1 signer, that the current user is the signer
and that the authentication method is not set to itsme, a new button
Download appears in the dialog window which is shown when clicking to
Sign a document. By clicking this new button everything happens under
the hood and the (already filled) document is downloaded right away.

Furthermore, after the download of the document, the view is changed
to the general view of the model we started from.

In the normal case where we go into the Sign app, open a template and
use the download button from there we are redirected to the kanban view
to the templates. In other flows like the one where we request a
signature from a sales order, we are redirected to the sale order form
view.

In this PR, there are also some more minor changes:

- The text in the buttons used to sign is changed from Sign Now to Sign
- The description present when adopting a signature or initials is
  changed slightly to adapt to the text present in the buttons of the
  different views where it is displayed: "By clicking Adopt & Sign," has
  now become "By signing,"
- We added a tip on the new Download button explaining its function.

closes odoo/enterprise#91318

Task: 4858331
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>