[FIX] sign: download documents from request dropdown component
As click event on DropdownItem rendered as `<a href="">` are prevented,
the recent refactoring (PR [1]) of the SignRequestDocumentsDropdown
component to use OWL components instead of Bootstrap JS introduce a
regression, preventing the user from downloading the documents.

This commit fixes it by properly using the dedicated download action
(i.e. `act_url` action with a `target="download"` attribute) when
selecting one of the dropdown's items.

Steps to reproduce:
- Open Sign
- In template, choose one and click Send
- Choose your user as Customer and click Send
- Go to Documents menu > My Documents > open the sent document
- Open Download dropdown and click on one of the item
=> download should happen but doesn't

[1]: https://github.com/odoo/enterprise/pull/89727

closes odoo/enterprise#91328

Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>