[FIX] account_accountant_batch_payment: only reconcile unreconciled payment term lines
Steps to reproduce:
- Install accountant

- Go to "Accounting / Configuration / Invoicing / Payment Terms"
- Create a payment term having at least 2 installments (e.g. 20%-80%)

- Create a bill:
  * Vendor: [any]
  * Payment terms: [the created payment term]
  * Amount: 1000
- Confirm the bill

- Pay the first installment from the bill (i.e. 200)
- Go to "Accounting / Vendors / Payments"
- Select the payment and create a batch
- Validate the batch payment

- From Accounting dashboard, go to Bank journal
- Create a statement:
  * Partner: [None]
  * Amount: -200
- Reconcile the statement with the batch payment

- From the bill, pay the second installment (i.e. 800)
- Go to "Accounting / Vendors / Payments"
- Select the payment and create a batch
- Validate the batch payment

- From Accounting dashboard, go to Bank journal
- Create a statement:
  * Partner: [None]
  * Amount: -800
- Reconcile the statement with the batch payment

Issue:
The following UserError is raised:
"You are trying to reconcile some entries that are already reconciled."

Cause:
When reconciling a batch payment, all the payment term lines from the invoices/bills
linked to the payment are used.
However, in this case, when the first batch payment has been reconciled, the first
payment term line of the bill (linked to the first installment) has been reconciled.
It should not be reused when reconciling the second batch payment.

Solution:
Exclude reconciled payment term lines when reconciling a batch payment.

opw-4830185

closes odoo/enterprise#94426

X-original-commit: 1622913fc2ed5e8c61f70869389a3117279f81d9
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>