[FIX] web_studio: Allow users to edit many2many_tags
Example of steps:
- Install contacts, web_studio and accounting
- Go to contacts, accounting tabs
- Open web_studio and click on Banks field
- Traceback

The error stems from the fact that we are using a widget in an o2m field
that uses a placeholder that is not defined for o2m.

One solution is to add the placeholder attribute for o2m in addition to
m2m and m2o.

opw-4937447

closes odoo/enterprise#91435

X-original-commit: 2d2beea2ab7fac19dd8a6581eab9ce22b71770e2
Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>
Signed-off-by: Achraf Ben Azzouz (abz) <abz@odoo.com>