[FIX] sale_renting_planning: make sure shifts generated by RO are published
Before this commit, when a shift is generated during the confirmation of
a Rental Order with a plannable product, if the resource assigned to the
shift generated is not a material one then the shift is not
automatically published.

This commit makes sure any shifts generated by a Rental Order is
automatically published.

task-4916537

X-original-commit: 36838e91c3d5567ae38426b63e1e11192654425c
Part-of: odoo/enterprise#92260
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>