[IMP] sale_commission: ensure rw cursor for report
Since https://github.com/odoo/enterprise/pull/88646 a temporary table is
created to speed up computation of commission report. As displaying the
report is creating and dropping the temporary table, it is not
compatible anymore with a read only cursor.

This commit ensures that _table_query don't create the temporary table
and that several methods that are readonly everywhere else work only
with a RW cursor:
 - web_search_read
 - formatted_read_group

task-4982555

closes odoo/enterprise#91352

closes odoo/enterprise#93425

Signed-off-by: Julien Castiaux (juc) <juc@odoo.com>
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>