[IMP] knowledge: add foldable section embedded component
This commit adds a new command block to Knowledge that enables the creation
of foldable sections. With the new /fold command, users can insert text
into these sections and toggle their visibility with a simple toggle button,
making it easy to show or hide content as needed. This feature will be
widely used in the upcoming accountant_knowledge release to support the
creation of audit reports within Knowledge.

Task-4840940

Part-of: odoo/enterprise#87928
Related: odoo/odoo#214712
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>