[FIX] l10n_hk_hr_payroll: LSP/SP
Fixes the computation of the Long Service Pay (LSP) and
Severance Pay (SP) by rewriting it to better comply with
the Hong Kong Employment Ordinance.

This also puts the bases of what will be needed to implement
the missing feature of MPF Offsetting. While this mechanism
has been abolished, it can still be applied by employers for
the period of employment up to the abolishment date.

task-4984107

Part-of: odoo/enterprise#91508
Related: odoo/odoo#221510
Related: odoo/upgrade#8159
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>