[IMP] sign: fix naming of signers
Currently, the sidebar of a sign template has an internal id which
starts at 0 and is used as an id when you make an orm call to
create a new sign.role.item record (a signer on the template). Until
now, the id is also used to create the name of the signer, as
"Signer {id}" but this is problematic if you eliminate signers and then
create new ones because it leaves holes in the signer naming. The
solution is to always use the length of the array containing the
currently assigned signers as the number to insert into the name, since
the new signer is always added at the end. Furthermore, when a signer is
deleted, we should compute the names of the remaining signers to match
their position in the array containing the assigned signers.

For example, if we create 4 signers, Signer {1,2,3,4} and then delete
Signer 3, we would like Signer 4 to be renamed Signer 3, since the next
added signer will be named Signer 4 according to the length of the array
containing them. A small caveat is that we don't want to rename already
renamed Signers, because we don't want to lose personalization.

This is the objective of this PR.

closes odoo/enterprise#91434

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>