[FIX] account_accountant: fix multi_currency payment_widget
To reproduce:
- Make a statement line for partner_a for 200 $
- Make an invoice for same partner for 100 $
- Reconcile it with the statement line
- Duplicate the invoice and post it
=> First issue, you don't see the 100$ left on the statement
line in the payment widget of the invoice.
While it's still reconciliable

To fix that, we only remove fully reconciled statement lines.

Second issue:
- Have a bank journal in EUR  with a rate of 2
- Make a statement line for 400€ in this journal
- Create an invoice for the same partner of 100$
=> The widget proposes a statement line of 400$ instead of a 200$
equivalent (with the rate conversion)

The issue is that we convert with a foreign_currency_id that is not
present in that case.

To fix both, we change the way we compute the amount to always use
the residual converted from the right currency

closes odoo/enterprise#92873

X-original-commit: bbc2072ce84047590798945d55d9f8823826bd32
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>