[IMP] l10n_mx_edi_landing: Improved customs workflow
Previous this commit, the customs workflow was somewhat of a mess. Currently
we depended of the origin movements linked to the valuation layer created
when delivering stock to a customer to be able to compute the customs numbers
to be invoiced. This caused issues for tracking and knowing what and how many
is being invoiced and sent to the SAT with the CFDI. This issue is magnified
if customer uses returns which breaks the whole original workflow.

From this commit this is now reworked so that user can select what is invoiced
using Lot/SN(s) delivered.

The user can link a landed cost and its customs number to a lot by selecting
the transfer or manually selecting the landed cost on the lot

An option is added at product level to be able to select the lots and the quantity
on an invoice line

target: master
task-3731106

closes odoo/enterprise#91745

Related: odoo/odoo#225271
Related: odoo/upgrade#8179
Signed-off-by: Laurent Smet (las) <las@odoo.com>