[FIX] account_accountant: reconciliation rounding lost precision case
The aim of this commit is to keep the original foreign amount currency
if the computation ends up close enough to it.

Before this commit:
The reconciliation process was losing so much precision that it could mess up
the reconciliation of one single invoice.

After this commit:
We keep the original amount as it is most probably the correct one.

Context:
With a rate of 1 US$ = 5.421327349 R$ and an invoice of 143.62 R$, we
convert the amount in US$ which is 26.491668921649627 US$.
As we have to round it for the accounting, we end up with 26.49 US$ as
company currency amount, losing the rest of the decimals.
During the reconciliation process, we convert back the US$ to R$ ending
up with 143.61096147501 R$ that have to be rounded to 143.61 R$.

This creates a difference of 0.01 R$ which surfaces later on.

Chosen solution:
As we still have the original currency amount and the rate, we are able
to recompute the raw numbers and we are able to make "fairer" comparison
between the amounts.
If we can confidently tell that the amounts are close enough, we
can just keep the original amount and prevent all those rounding errors
to be taken into account.

opw-4937508

closes odoo/enterprise#91462

closes odoo/enterprise#91857

X-original-commit: 7622471cfa3c77997d4c25d8bd0f1ce6dcfa9e5a
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>