[FIX] documents: fix updating search panel names
... after changing from the details panel.

Something changed in 18.4 (fw or documents) such that we
* sometimes we didn't trigger a reload of the search panel
  after modifying a folder
* sometimes the reload of the searchpanel was triggered
 *before* the initiation (and return) of the write call,
 so the search panel was updated with the old values

Note that this asynchronous misbehavior was not easily
visible in hoot tests, but the call count enables to
highlight the problem (and drive a suitable solution).

The fix is actually an opportunity to keep reloading the
search panel on update inside `update` (i.e., saving) flows.

Also last style cleanups in documents_details_panel.js.

closes odoo/enterprise#91866

X-original-commit: 9601c42069537f3859838a0517584d90155cdd40
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>