[IMP] l10n_hk_hr_payroll: ADW as a field
Extract the logic of the average daily wage
in a new computed field, allowing to display
it on the payslip.
This is an essential part of Hong Kong wages
calculation, affecting the amounts on the worked
days line, making it valuable to allow users to
see the amount used during calculations.

task-4971776

Part-of: odoo/enterprise#91508
Related: odoo/odoo#221510
Related: odoo/upgrade#8159
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>