[IMP] l10n_hk_hr_payroll: end of year pay improvement
Currently, the end of year payment is fixed to always
be included in december, for all users.
This is wrong on multiple levels;
- This is not a mandatory advantage in HK, and companies
can choose not to provide it.
- Some companies will give this advantage in january, or for
the lunar new year instead.
- Or some will even give it as a discretionary bonus for top
performers/... and not to all employees.

We add some improvement to the system by:
- Adding a setting to select the month of the payment.
- It can optionally be left empty to never automatically add it.
Sometimes, you may want to provide it as a discretionary bonus.

task-4982154

Part-of: odoo/enterprise#91508
Related: odoo/odoo#221510
Related: odoo/upgrade#8159
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>