[IMP] l10n_hk_hr_payroll: Fix of computation 713
As of now, when calculating 80% work entries using
the "713" ordinance, we are taking the max between
the Average Daily Wages and the Daily Wages of the
month.

The law states that the ADW * 80% should be used for
the computation.
It also states that any term in a contract that attempts
to reduce or take away a right given by the Ordinance is void.
It does not prohibit to offering terms that are more favourable.

This is why the current logic is in place, but when in doubt, we
should be following the law as closely as we can, which means that
we should remove that logic and strictly use the ADW.

task-4982025

Part-of: odoo/enterprise#91508
Related: odoo/odoo#221510
Related: odoo/upgrade#8159
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>