[IMP] ai: trigger AI prompt onchange on mutations
Purpose
=======
In Documents, there's a warning when we don't insert record in the
folder's prompt. But at the moment, that warning is displayed only
when we blur the component. We want it to be visible / hidden as
soon as possible.

Task-4915266

Part-of: odoo/enterprise#90124
Related: odoo/odoo#218725
Related: odoo/upgrade#8073
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>