[IMP] stock, *: allow multi-level packages
Before this change, it wasn't possible to handle multiple levels of
packages within Odoo. One simple example of this is having a few boxes
put on top of a pallet, which is a *very* common occurence in stock.

Changes summary:
- Packages inside other packages
- Rework of the put in pack, can display a widget if enabled in
  Operation Type
- You can move a package containing other packages, this will move all
of them
- The handling of a package is *always* done on the move lines directly
- Packages now holds several more information regarding their current
reservation
- A new `stock.package.history` record is added to keep track of what
  was moved in previous transfers, as package hierarchy may be different
  between the past and now

Task-4314600

Part-of: odoo/enterprise#82453
Related: odoo/odoo#203987
Related: odoo/upgrade#7576
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>