[FIX] stock_barcode: fix line addQuantity button issue
Steps to reproduce:
- Install Purchase, Stock and Barcode apps
- Create a product tracked by serial number
- Create a PO of this product and confirm it (demand 3)
- Open Barcode application for the PO's receipt
- Start entering serial numbers for the 3 products
- For each line, use the decrement button and remove all 3 lines
- Make sure the last removed line is not the first scanned serial
- Re-enter the serial number of the last removed line

Issue:
The parent line of the 3 sublines has the lot_name fixed on the
first scanned serial_number. So when decrementing the last_line,
it will always use the virtual_id of the first serial_number
scanned, leading to decrementing the wrong serial number. It
starts showing that the next scanned number is already used,
because it wasn't properly removed in the first place.

opw-4646765

closes odoo/enterprise#91867

X-original-commit: ec48997251d93e31ef95079c1fc5873df8849dda
Signed-off-by: Steve Van Essche <svs@odoo.com>