[IMP] sale_planning: group by resource only for 'To plan' action
Before this commit, when SO, contained a product configured to generate
shifts, is confirmed, a `To Plan` stat button is displayed in the form
view of that button. The problem is when the user clicks on that button,
he will go to gantt view of shifts grouped by role > resource and he
will not be able to plan any shifts.

This commit makes sure the user can plan the shifts generated inside
that gantt view. To do that, the default group by will be `Resource`.

closes odoo/enterprise#91602

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>