[IMP] l10n_be_hr_contract_salary: Auto-fill car on offer creation from contract template or version
In this PR, we fixed the assignation of the car in the contract when the contract template for the version has company car.
When creating an offer, automatically set the transportation car based on the linked contract template.
If the contract template has no car, fall back to the car defined on the employee version.
If neither has a car, leave the field empty.
Related task: 4926053.

closes odoo/enterprise#91609

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>