[FIX] accountant_knowledge: init currency table when needed
The method for generating the PDF of an audit report currently crashes
when computing the template variables "revenue" and "net accounting result".

This issue arises because the underlying method used to compute these
values runs raw SQL queries that perform a JOIN operation on the
`account_currency_table`. However, this table is not created automatically
by the system: It must be explicitly initialized by calling the method
`_init_currency_table`.

This commit updates the code to ensure that `_init_currency_table` is
called before invoking the methods that compute the total revenue, the
net accounting result, etc. This should prevent the crash and allow the
template variables to be computed correctly.

Steps to reproduce the issue on the runbot:
1. Open the Accounting App
2. In the navbar, click on "Accounting" > "Audit Reports"
3. Create a new audit report
4. Fill the form and save
5. Click on the "Print" button of the kanban card.

=> A new page should open displaying an internal error.

In the log of the runbot, we have the following log:

```
File "/data/build/enterprise/account_reports/models/account_report.py",
line 3899, in _compute_formula_batch_with_engine_domain
    self.env.cr.execute(query)
  File "/data/build/odoo/odoo/sql_db.py", line 426, in execute
    self._obj.execute(query, params)
psycopg2.errors.UndefinedTable: relation "account_currency_table" does not exist
LINE 10:                     LEFT JOIN account_currency_table
```

Task-4840940

closes odoo/enterprise#91625

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>