[IMP] account_accountant: match amount in between in negative
Before this commit, when having the match amount in between with negative amount
we had the min and max amount. Some user were confused that since the amount are
negative the min and max need to be inverted.
To solve that we rewrite the way the match amount were dealt to use a case when
that will enable to have an in between that works in both ways (min, max and
max, min)

notask-id

closes odoo/enterprise#92285

X-original-commit: c785090d6aee05e53963c9a64ee3078fccfd6a56
Related: odoo/odoo#222775
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>