[FIX] stock_barcode: Demo message not disappearing
Fixes demo message removal on barcode landing page. If
clicking "Keep it" the demo message now stays in UI,
and if clicking "Remove it", the message now disappear
(except in one edge case, detailed below).

AS-IS: Clicking "X" removes demo message before the confirm
or cancel of pop up clicked. If "Keep it" is clicked the
message doesn't reappear in current view and if clicking
"Remove it" it reappears after a page reload (because
display = false is set in action). The message finally
disappears on barcode open from dashboard.

NOW: The message only disappear if the user confirms with "Remove
it" in the pop-up, and it won't reappear unless edge case. If
clicking "Keep It" the message doesn't disapear.

Edge case: Because the action with the displayDemoMessage = false
is set on the server, it won't propagate to the client action until
a page reload. So if after discarding the message, if the user navigates
back to dashboard and back to Barcode -> message re-appears. A hacky
solution to override client action is not worth the effort

Sidenote, this message only appears on DBs with demo data.

task #4729518

closes odoo/enterprise#87450

Related: odoo/odoo#213852
Signed-off-by: Steve Van Essche <svs@odoo.com>