[FIX] web_studio: allow edition of the group_no_one
Steps to reproduce
==================

- In debug mode
- Go to Contacts
- Open any record
- Open studio
- Click on any field
- Add the "Extra rights / Technical Features" group to "Allow visibility to groups"

=> The group is not displayed as a tag

- Restore the view
- Now without debug mode (?debug=0), repeat the same steps

=> The field disappears, we need to toggle "Show Invisible Elements" to
see it again

Cause of the issue
==================

https://github.com/odoo/odoo/pull/179354/commits/15aeaf88c268f047b8b83a5aa66f5da246c2b675

When calling get_views, the "base.group_no_one" is removed from the
groups attribute and "invisible" is set to true if we are in debug mode.

Solution
========

Adding "base.group_no_one" is still the way to have the expected
behavior for now. We override some ir.ui.view functions when in studio
to be able to edit it.

According to the docstring of _postprocess_debug_to_cache, this feature
is temporary. Another solution will be needed in the future.

opw-4969262

closes odoo/enterprise#93663

X-original-commit: 9cea2ef130e2749c676b8f209c8d087451414ba0
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Hubert Van De Walle <huvw@odoo.com>