[FIX] account_loans: Fix error while searching Loan
Behaviour before this commit:
While searching for 'Loan' in the Loan Analysis view, a traceback
appears since the filter_domain for the search is on 'name' but
'name' is not present in account.loan.line, instead it is
loan_name.

Desired behaviour after this commit:
The filter_domain for 'Loan' is updated to search for loan_name
instead of loan, and domain for 'display_name' is removed.

TaskID:4991369

closes odoo/enterprise#91796

X-original-commit: 80e9e9f6beb474255d379a45634d5cd1b8d0054a
Signed-off-by: Nicolas Viseur (vin) <vin@odoo.com>
Signed-off-by: Manav Shah (sham) <sham@odoo.com>