[FIX] esg: fix daterange widget inheritance
The widgets `esgDateRangeField` and `esgListDateRangeField` inherit respectively from `dateTimeField` and `listDateTimeField`.
As their name implies, those widgets are made to handle datetime fields, whereas the esg widgets are used on dates.
While it does not seem to make a functional difference, it produces a warning in the browser console.
This commit makes the esg widgets inherit from the "date" counterpart instead.

closes odoo/enterprise#91666

X-original-commit: a457168342254d4e383245f452f784bbb3bda583
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>