[FIX] accountant_knowledge: provide data-oe-id to embedded account report
Currently, the embedded account reports included in templates are missing
the `data-oe-id` attribute. This attribute is essential because it allows
us to uniquely identify and target a specific report element in order to
update its associated options. (See: `update_embedded_audit_report_options`)

Without this attribute, we are unable to locate and update the corresponding
element in the DOM when the user customizes the report. As a result, users
are unable to save changes to the report options for embedded reports
originating from templates.

Steps to reproduce the issue:
1. Create an account report
2. Open the "Balance Sheet" article
3. Click on the "Customize" button of the embedded balance sheet
4. Change the options (in the control panel)
5. Go back to the article
6. Re-Click on the "Customize" button

=> Changes have not been saved

TO BE: When a user clicks the "Customize" button of an embedded account
report and modifies the options, those changes should be saved.

To resolve this issue, we will add a `data-oe-id` attribute to the
embedded account reports defined in the templates. With those changes,
users will be able to modify and save the options of audit reports
embedded in templates.

Task-4840940

closes odoo/enterprise#91652

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>