[IMP] iot: add tests for iot_http_service
This commit brings unit tests for the iot http service, and small
fixes to it.

It also brings a tour to ensure using 'Test' print on the iot device
record uses websocket (after failing using webrtc and longpolling),
and receives a response in the right format when calling the callback
controller (that the iot box calls to send a websocket message).

closes odoo/enterprise#91653

Signed-off-by: Max Whale (mawh) <mawh@odoo.com>