[FIX] account_reports: Maintain correct active company context when generating XML
Error: When generating a tax return, there is a bug that can omit certain company data from the generated XML. This happens when a child company is earlier alphabetically than the parent company. The company ids are ordered alphabetically for display purposes on the tax return view, however when writing this value into the allowed_company_ids context value, the first id in the list will override the current company environment variable during xml generation.

To fix this we pass the active company into the context to make sure it is not overwritten.

OPW-4964467

closes odoo/enterprise#92740

X-original-commit: f273d9d7340da773d56c4ce16df8cddd12068ad8
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>