[FIX] sale_subscription : Link payment to invoice when subscription
Steps to reproduce:
	- Activate a payment method
	- Create a subscription and preview it in portal view
	- Pay this order through the order's portal view
	- Navigate to the invoice got generated
	- Notice there is not 'Payments' smart button

Cause:
Since https://github.com/odoo/enterprise/pull/77432/commits/02f72f94b7c996fa19c6bfc70a704d27aa677245 we stopped creating the invoice before
the payment and the transaction in the controller and we let
_post_process do this invoice creation. So the invoice actually
gets created after the payment not before and when preparing
the payment values the invoice_ids will be False

https://github.com/odoo/odoo/blob/6c88e6ffd2c906b72b31d3c34ad3d6da5efba204/addons/account_payment/models/payment_transaction.py#L155-L170

Then we create the invoice for the transaction and we don't
propagate it to the payment.

https://github.com/odoo/enterprise/blob/20fd574367757c06f30c344056e8b8d91ba8244c/sale_subscription/models/payment_transaction.py#L72-L84

Fix:
Link the payment to the invoice after creating the invoice
and link it to the transaction.

opw-4867631

closes odoo/enterprise#92579

X-original-commit: 9368ff378fea1acd03bb8e627ad635c79d97a81c
Signed-off-by: Youssef Bashandy (yoba) <yoba@odoo.com>