[FIX] l10n_co_edi: prevent traceback in website
To reproduce:
=============
1- Install Colombia accounting
2- Install eCommerce
3- Set company address to Colombia
4- Go to website and add item to cart
5- Checkout & click "Add new address"
→ Traceback occurs

Problem & Solution:
===================
The field l10n_latam_identification_type_id is always present in the
address form but its HTML element changes: it's rendered as a `select`
when is_commercial_address and can_edit_vat is true, and as an `input`
otherwise. The JS code only looked for the `select`, so when the `input`
was rendered instead, the selector was null and caused a traceback.
Update the selector to look for the field in either select or input form,
so the code works regardless of the rendering conditions.

useful-links:
https://github.com/odoo/odoo/blob/4806b08dcfc965cdbd463269be078f8da9f48863/addons/l10n_latam_base/views/portal_address_templates.xml#L10-L11
https://github.com/odoo/odoo/blob/4806b08dcfc965cdbd463269be078f8da9f48863/addons/l10n_latam_base/views/portal_address_templates.xml#L32-L33
opw-4969089

closes odoo/enterprise#92335

X-original-commit: 5b882d9e97ff4862bb738f85dc7cb452de1315c2
Signed-off-by: Ali Alfie (alal) <alal@odoo.com>
Signed-off-by: Saif Allah Ben Khalil (sben) <sben@odoo.com>